password = username?!
When a user changes his/her own password, it did not set it to the new password. No, it set the password to the user's username. This eliminated 1 step in the two-step verification for hackers, and successfully prevented the user from being able to ever login again unless they knew about the bug or asked an admin to reset the password again.